cluster/dht: suppress file migration error for node not supposed to migrate file
A rebalance process does a lookup for every file in the dir it is processing
before checking if it supposed to migrate the file.
In this issue there are two rebalance processses running on a replica subvol:
R1 is migrating the FILE.
R2 is not supposed to migrate the FILE, but it does a lookup and
   finds a stale linkfile which is mostly due to a stale layout.
   Then, it tries to unlink the stale linkfile and gets EBUSY
   as the linkfile fd is open due R1 migration.
   As a result a misleading error msg about FILE migration failure
   due EBUSY is logged in R2 logfile.

Fix:
suppress the error in case it occured in a node that
is not supposed to migrate the file.

fixes: #1371
Change-Id: I37832b404e2b0cc40ac5caf45f14c32c891e71f3
Signed-off-by: Tamar Shacked <tshacked@redhat.com>