libglusterfs/xlator: undefined symbol xlator_api
Issue:
On executiing the command gluster vol set help, an error
comes up in glusterd logs stating `undefined symbol: xlator_api`.
This issue is seen for the rpc-transport/socket.so file.

Fix:
The symbol `xlator_api` is not found in rpc-transport/socket.so
file as it is not a xlator but a shared object for transport.
In the `xlator.c` file, there is a function `xlator_volopt_dynload`,
which looks for the default values of the options available in gluster,
which is stored inside the respective xlator files for different voltypes.
In each of these files the `options` object is present which contains
the default values, which is therefore referenced from the `options` data
member of `xlator_api` object in case of xlators.But, since
`rpc-transport/socket.so` is not an xlator we don't have the `xlator_api`
object present to point to that object. So, in case of
`rpc-transport/socket.so` type we are accesing the `options` object
directly from the `xlator_volopt_dynload` function to fetch the default
values for the available options.

Fixes: #827
Change-Id: I3b2b0c1f2a11896be250aaca1a33a65b044991d5
Signed-off-by: nik-redhat <nladha@redhat.com>